requirements.txt
psycopg2-binary
sqlalchemy
pandas
streamlit
python-dotenv
